requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
Pillow>=9.5.0
gunicorn>=20.1.0
dj-database-url>=1.2.0
whitenoise>=6.5.0
